refactor: restructure top-level firmware initialization
