+++ firmware/firmware.cpp
@@ -1,15 +1,20 @@
-#include "main.h"
-
-#include <Arduino.h>
 #include "openhaptics.h"
+
+#if defined(ARDUINO)
+#include <Arduino.h>
+#endif // ARDUINO
 
 #if defined(SERIAL_PLOTTER) && SERIAL_PLOTTER == true
 #include <components/serial_plotter.hpp>
-#endif
+#endif // SERIAL_PLOTTER
 
 #ifndef PIO_UNIT_TESTING
 
+OpenHaptics App;
+
 extern void setupMode(void);
+
+#if defined(ARDUINO)
 
 void setup() {
   Serial.begin(115200);
@@ -18,7 +23,7 @@ void setup() {
 #if defined(SERIAL_PLOTTER) && SERIAL_PLOTTER == true
   auto* serialOutputState = new OH::SerialPlotter_OutputStates<HardwareSerial>(Serial, App.getOutput());
   App.registerComponent(serialOutputState);
-#endif
+#endif // SERIAL_PLOTTER
 
   App.begin();
 
@@ -28,4 +33,6 @@ void setup() {
 
 void loop() {}
 
-#endif
+#endif // ARDUINO
+
+#endif // PIO_UNIT_TESTING
