+++ firmware/mode_configs/bhaptics/tactal.cpp
@@ -4,39 +4,33 @@
 #include <Arduino.h>
 #include <Wire.h>
 
-#include <utility.hpp>
-
 #include "openhaptics.h"
 
 #include <connection_bhble.hpp>
 #include <output_writers/pwm.hpp>
 
 #if defined(BATTERY_ENABLED) && BATTERY_ENABLED == true
-#include <battery/adc_naive.hpp>
+#include "battery/adc_naive.hpp"
 #endif
 
 using namespace OH;
 using namespace BH;
 
+extern OpenHaptics App;
+
 #pragma region bHaptics_trash
 
-const uint16_t _bh_size_x = 6;
-const uint16_t _bh_size_y = 1;
-
-inline oh_output_point_t* make_point(oh_output_coord_t x, oh_output_coord_t y) {
-  return getPoint(x, y, (oh_output_coord_t) (_bh_size_x - 1), (oh_output_coord_t) (_bh_size_y - 1));
-}
-
-oh_output_point_t* indexesToPoints[_bh_size_x * _bh_size_y] = {
-    make_point(0, 0), make_point(1, 0), make_point(2, 0),
-    make_point(3, 0), make_point(4, 0), make_point(5, 0)};
+static const oh_output_point_t* indexesToPoints[BH_LAYOUT_TACTAL_SIZE] = BH_LAYOUT_TACTAL;
 
 void vestMotorTransformer(std::string& value) {
-  for (size_t i = 0; i < _bh_size_x; i++) {
+  for (size_t i = 0; i < BH_LAYOUT_TACTAL_SIZE; i++) {
     uint8_t byte = value[i];
-    oh_output_data_t output_0;
-    output_0.point = *indexesToPoints[i];
-    output_0.intensity = map(byte, 0, 100, 0, UINT16_MAX);
+
+    oh_output_data_t output_0{
+      .point = *indexesToPoints[i],
+      .intensity = static_cast<oh_output_intensity_t>(map(byte, 0, 100, 0, OH_OUTPUT_INTENSITY_MAX)),
+    };
+
     App.getOutput()->writeOutput(OUTPUT_PATH_ACCESSORY, output_0);
   }
 }
@@ -44,18 +38,18 @@ void vestMotorTransformer(std::string& value) {
 #pragma endregion bHaptics_trash
 
 void setupMode() {
-  // Configure PWM pins to their positions on the glove
-  auto gloveOutputs = mapMatrixCoordinates<AbstractOutputWriter>({
+  // Configure PWM pins to their positions on the face
+  auto faceOutputs = mapMatrixCoordinates<AbstractOutputWriter>({
       // clang-format off
       {new PWMOutputWriter(32), new PWMOutputWriter(33), new PWMOutputWriter(25), new PWMOutputWriter(26), new PWMOutputWriter(27), new PWMOutputWriter(14)},
       // clang-format on
   });
 
-  OutputComponent* glove = new ClosestOutputComponent(gloveOutputs);
-  App.getOutput()->addComponent(OUTPUT_PATH_ACCESSORY, glove);
+  OutputComponent* face = new ClosestOutputComponent(OUTPUT_PATH_ACCESSORY, faceOutputs);
+  App.getOutput()->addComponent(face);
 
 #if defined(BATTERY_ENABLED) && BATTERY_ENABLED == true
-  AbstractBattery* battery = new ADCNaiveBattery(33, { .sampleRate = BATTERY_SAMPLE_RATE }, &App);
+  AbstractBattery* battery = new ADCNaiveBattery(33, { .sampleRate = BATTERY_SAMPLE_RATE }, &App, tskNO_AFFINITY);
   App.setBattery(battery);
 #endif
