+++ firmware/mode_configs/bhaptics/tactosy2.cpp
@@ -16,12 +16,14 @@
 using namespace OH;
 using namespace BH;
 
+extern OpenHaptics App;
+
 #pragma region bHaptics_trash
 
-oh_output_point_t* indexesToPoints[BH_LAYOUT_TACTOSYH_SIZE] = BH_LAYOUT_TACTOSYH;
+oh_output_point_t* indexesToPoints[BH_LAYOUT_TACTOSY2_SIZE] = BH_LAYOUT_TACTOSY2;
 
 void vestMotorTransformer(std::string& value) {
-  for (size_t i = 0; i < BH_LAYOUT_TACTOSYH_SIZE; i++) {
+  for (size_t i = 0; i < BH_LAYOUT_TACTOSY2_SIZE; i++) {
     uint8_t byte = value[i];
 
     oh_output_data_t output_0{
@@ -36,17 +38,16 @@ void vestMotorTransformer(std::string& value) {
 #pragma endregion bHaptics_trash
 
 void setupMode() {
-  // Configure PWM pins to their positions on the hands
-  auto handOutputs = mapMatrixCoordinates<AbstractOutputWriter>({
+  // Configure PWM pins to their positions on the forearm
+  auto forearmOutputs = mapMatrixCoordinates<AbstractOutputWriter>({
       // clang-format off
-      {new PWMOutputWriter(32)},
-      {new PWMOutputWriter(33)},
-      {new PWMOutputWriter(25)}
+      {new PWMOutputWriter(32), new PWMOutputWriter(33), new PWMOutputWriter(25)},
+      {new PWMOutputWriter(26), new PWMOutputWriter(27), new PWMOutputWriter(14)},
       // clang-format on
   });
 
-  OutputComponent* hand = new ClosestOutputComponent(OUTPUT_PATH_ACCESSORY, handOutputs);
-  App.getOutput()->addComponent(hand);
+  OutputComponent* forearm = new ClosestOutputComponent(OUTPUT_PATH_ACCESSORY, forearmOutputs);
+  App.getOutput()->addComponent(forearm);
 
 #if defined(BATTERY_ENABLED) && BATTERY_ENABLED == true
   AbstractBattery* battery = new ADCNaiveBattery(33, { .sampleRate = BATTERY_SAMPLE_RATE }, &App, tskNO_AFFINITY);
