+++ firmware/mode_configs/bhaptics/tactosyh.cpp
@@ -16,13 +16,15 @@
 using namespace OH;
 using namespace BH;
 
+extern OpenHaptics App;
+
 #pragma region bHaptics_trash
 
-oh_output_point_t* indexesToPoints[BH_LAYOUT_TACTOSYF_SIZE] = BH_LAYOUT_TACTOSYF;
+oh_output_point_t* indexesToPoints[BH_LAYOUT_TACTOSYH_SIZE] = BH_LAYOUT_TACTOSYH;
 
 void vestMotorTransformer(std::string& value) {
-  for (size_t i = 0; i < BH_LAYOUT_TACTOSYF_SIZE; i++) {
-  uint8_t byte = value[i];
+  for (size_t i = 0; i < BH_LAYOUT_TACTOSYH_SIZE; i++) {
+    uint8_t byte = value[i];
 
     oh_output_data_t output_0{
       .point = *indexesToPoints[i],
@@ -36,15 +38,17 @@ void vestMotorTransformer(std::string& value) {
 #pragma endregion bHaptics_trash
 
 void setupMode() {
-  // Configure PWM pins to their positions on the feet
-  auto footOutputs = mapMatrixCoordinates<AbstractOutputWriter>({
+  // Configure PWM pins to their positions on the hands
+  auto handOutputs = mapMatrixCoordinates<AbstractOutputWriter>({
       // clang-format off
-      {new PWMOutputWriter(32), new PWMOutputWriter(33), new PWMOutputWriter(25)}
+      {new PWMOutputWriter(32)},
+      {new PWMOutputWriter(33)},
+      {new PWMOutputWriter(25)}
       // clang-format on
   });
 
-  OutputComponent* foot = new ClosestOutputComponent(OUTPUT_PATH_ACCESSORY, footOutputs);
-  App.getOutput()->addComponent(foot);
+  OutputComponent* hand = new ClosestOutputComponent(OUTPUT_PATH_ACCESSORY, handOutputs);
+  App.getOutput()->addComponent(hand);
 
 #if defined(BATTERY_ENABLED) && BATTERY_ENABLED == true
   AbstractBattery* battery = new ADCNaiveBattery(33, { .sampleRate = BATTERY_SAMPLE_RATE }, &App, tskNO_AFFINITY);
