+++ firmware/mode_configs/bhaptics/tactsuit_x40.cpp
@@ -17,6 +17,8 @@
 using namespace OH;
 using namespace BH;
 
+extern OpenHaptics App;
+
 #pragma region bHaptics_trash
 
 oh_output_point_t* indexesToPoints[BH_LAYOUT_TACTSUITX40_SIZE] = BH_LAYOUT_TACTSUITX40;
