--- include/main.h
@@ -1,2 +1,0 @@
-#pragma once
-#include "config/all.h"
