+++ ini/bhaptics.ini
@@ -15,7 +15,7 @@ build_flags = ${bhaptics.build_flags}
 	'-D BH_SERIAL_NUMBER={ 0x0d, 0x3a, 0xeb, 0x77, 0xbe, 0xf8, 0x7a, 0x1e, 0x3b, 0x2a }'
 build_unflags = ${bhaptics.build_unflags}
 build_src_filter  = ${bhaptics.build_src_filter}
-	+<modes/bhaptics/tactsuit_x16.cpp>
+	+<mode_configs/bhaptics/tactsuit_x16.cpp>
 lib_deps    = ${bhaptics.lib_deps}
 
 [env:bhaptics_tactsuit_x16_pca9685]
@@ -26,7 +26,7 @@ build_flags = ${bhaptics.build_flags}
 	'-D BH_SERIAL_NUMBER={ 0x0d, 0x3a, 0xeb, 0x77, 0xbe, 0xf8, 0x7a, 0x1e, 0x3b, 0x2a }'
 build_unflags = ${bhaptics.build_unflags}
 build_src_filter  = ${bhaptics.build_src_filter}
-	+<modes/bhaptics/tactsuit_x16_pca9685.cpp>
+	+<mode_configs/bhaptics/tactsuit_x16_pca9685.cpp>
 lib_deps    = ${bhaptics.lib_deps}
 
 [env:bhaptics_tactsuit_x40]
@@ -37,7 +37,7 @@ build_flags = ${bhaptics.build_flags}
 	'-D BH_SERIAL_NUMBER={ 0xcf, 0xcb, 0x0d, 0x95, 0x5f, 0xf6, 0xee, 0x2c, 0xbd, 0x73 }'
 build_unflags = ${bhaptics.build_unflags}
 build_src_filter  = ${bhaptics.build_src_filter}
-	+<modes/bhaptics/tactsuit_x40.cpp>
+	+<mode_configs/bhaptics/tactsuit_x40.cpp>
 lib_deps    = ${bhaptics.lib_deps}
 
 [env:bhaptics_tactosy2_forearm_left]
@@ -48,7 +48,7 @@ build_flags = ${bhaptics.build_flags}
 	'-D BH_SERIAL_NUMBER={ 0xa0, 0xba, 0x0a, 0xd1, 0xbf, 0x36, 0x11, 0x30, 0xa4, 0xff }'
 build_unflags = ${bhaptics.build_unflags}
 build_src_filter  = ${bhaptics.build_src_filter}
-	+<modes/bhaptics/tactosy2.cpp>
+	+<mode_configs/bhaptics/tactosy2.cpp>
 lib_deps    = ${bhaptics.lib_deps}
 
 [env:bhaptics_tactosy2_forearm_right]
@@ -59,7 +59,7 @@ build_flags = ${bhaptics.build_flags}
 	'-D BH_SERIAL_NUMBER={ 0xb0, 0x1c, 0xc1, 0xf8, 0xec, 0x12, 0x18, 0x4e, 0x09, 0x77 }'
 build_unflags = ${bhaptics.build_unflags}
 build_src_filter  = ${bhaptics.build_src_filter}
-	+<modes/bhaptics/tactosy2.cpp>
+	+<mode_configs/bhaptics/tactosy2.cpp>
 lib_deps    = ${bhaptics.lib_deps}
 
 [env:bhaptics_tactosyh_hand_left]
@@ -70,7 +70,7 @@ build_flags = ${bhaptics.build_flags}
 	'-D BH_SERIAL_NUMBER={ 0xc1, 0x36, 0xdc, 0x21, 0xc9, 0xd4, 0x17, 0x85, 0xbb, 0x90 }'
 build_unflags = ${bhaptics.build_unflags}
 build_src_filter  = ${bhaptics.build_src_filter}
-	+<modes/bhaptics/tactosyh.cpp>
+	+<mode_configs/bhaptics/tactosyh.cpp>
 lib_deps    = ${bhaptics.lib_deps}
 
 [env:bhaptics_tactosyh_hand_right]
@@ -81,7 +81,7 @@ build_flags = ${bhaptics.build_flags}
 	'-D BH_SERIAL_NUMBER={ 0xc7, 0x5f, 0x3b, 0x06, 0x38, 0xba, 0x34, 0xfa, 0x36, 0xc1 }'
 build_unflags = ${bhaptics.build_unflags}
 build_src_filter  = ${bhaptics.build_src_filter}
-	+<modes/bhaptics/tactosyh.cpp>
+	+<mode_configs/bhaptics/tactosyh.cpp>
 lib_deps    = ${bhaptics.lib_deps}
 
 [env:bhaptics_tactosyf_foot_left]
@@ -92,7 +92,7 @@ build_flags = ${bhaptics.build_flags}
 	'-D BH_SERIAL_NUMBER={ 0x1a, 0x45, 0x83, 0x44, 0x03, 0xc5, 0xf3, 0xc3, 0xf3, 0xb8 }'
 build_unflags = ${bhaptics.build_unflags}
 build_src_filter  = ${bhaptics.build_src_filter}
-	+<modes/bhaptics/tactosyf.cpp>
+	+<mode_configs/bhaptics/tactosyf.cpp>
 lib_deps    = ${bhaptics.lib_deps}
 
 [env:bhaptics_tactosyf_foot_right]
@@ -103,7 +103,7 @@ build_flags = ${bhaptics.build_flags}
 	'-D BH_SERIAL_NUMBER={ 0x14, 0xb9, 0x02, 0x62, 0x41, 0xe4, 0x04, 0xb2, 0xc5, 0x11 }'
 build_unflags = ${bhaptics.build_unflags}
 build_src_filter  = ${bhaptics.build_src_filter}
-	+<modes/bhaptics/tactosyf.cpp>
+	+<mode_configs/bhaptics/tactosyf.cpp>
 lib_deps    = ${bhaptics.lib_deps}
 
 [env:bhaptics_tactal]
@@ -114,7 +114,7 @@ build_flags = ${bhaptics.build_flags}
 	'-D BH_SERIAL_NUMBER={ 0xed, 0xcb, 0x55, 0x7c, 0xd7, 0xb9, 0x16, 0xc5, 0x18, 0x2a }'
 build_unflags = ${bhaptics.build_unflags}
 build_src_filter  = ${bhaptics.build_src_filter}
-	+<modes/bhaptics/tactal.cpp>
+	+<mode_configs/bhaptics/tactal.cpp>
 lib_deps    = ${bhaptics.lib_deps}
 
 [env:bhaptics_tactglove_l]
@@ -125,7 +125,7 @@ build_flags = ${bhaptics.build_flags}
 	'-D BH_SERIAL_NUMBER={ 0xcd, 0x0b, 0x81, 0x45, 0x85, 0xf9, 0x2b, 0x6c, 0xed, 0x5b }'
 build_unflags = ${bhaptics.build_unflags}
 build_src_filter  = ${bhaptics.build_src_filter}
-	+<modes/bhaptics/tactglove.cpp>
+	+<mode_configs/bhaptics/tactglove.cpp>
 lib_deps    = ${bhaptics.lib_deps}
 
 [env:bhaptics_tactglove_r]
@@ -136,5 +136,5 @@ build_flags = ${bhaptics.build_flags}
 	'-D BH_SERIAL_NUMBER={ 0x12, 0x0b, 0xae, 0xbf, 0xbc, 0x90, 0x3b, 0x0d, 0x84, 0xdd }'
 build_unflags = ${bhaptics.build_unflags}
 build_src_filter  = ${bhaptics.build_src_filter}
-	+<modes/bhaptics/tactglove.cpp>
+	+<mode_configs/bhaptics/tactglove.cpp>
 lib_deps    = ${bhaptics.lib_deps}
