+++ platformio.ini
@@ -22,13 +22,13 @@ build_flags       =
 	-D CORE_DEBUG_LEVEL=3
 	-D DEBUG_MODE=0
 	-D DEBUG_ESP_PORT=Serial
-;	-D SERIAL_PLOTTER=true
+	-D SERIAL_PLOTTER=true
 	-D BATTERY_ENABLED=true
 build_unflags     =
 	-std=gnu++11
 build_src_filter  =
 	+<*>
-	-<modes>
+	-<mode_configs>
 lib_deps          =
 	Wire
 	SPI
@@ -55,4 +55,4 @@ test_ignore      = test_desktop
 build_flags      = ${common.build_flags}
 build_unflags    = ${common.build_unflags}
 build_src_filter = ${common.build_src_filter}
-	+<modes/test.cpp>
+	+<mode_configs/test.cpp>
